Move TmcClient into an external library
For testing purposes, move tmcclient to lib/tmcclient. This way it stays
decoupled from the primary plugin code and can be tested independently from
other plugin code.
