+++ lib/tmcclient/tmcclient.h
@@ -6,38 +6,40 @@
 #include <QNetworkRequest>
 #include <QNetworkReply>
 #include <QUrl>
-#include <QDateTime>
 #include <QFile>
 #include <QDebug>
 #include <QUrlQuery>
 
 #include "course.h"
-#include "downloadpanel.h"
 
 class TmcClient : public QObject
 {
     Q_OBJECT
 public:
     explicit TmcClient(QObject *parent = 0);
+
+    void setNetworkManager(QNetworkAccessManager *m);
     void authenticate(QString username, QString password, bool savePassword);
     void getUserInfo();
-    void getExerciseList(Course* course);
-    void getExerciseZip(Exercise *ex, DownloadPanel *downloadPanel);
-    Course * getCourse();
+    void getExerciseList(Course *course);
+    QNetworkReply* getExerciseZip(Exercise *ex);
+
 signals:
+    void TMCError(QString errorString);
     void loginFinished();
-    void exerciseListReady();
+    void exerciseListReady(Course *course);
     void exerciseZipReady(Exercise *ex);
 
 public slots:
     void authenticationFinished (QNetworkReply *reply);
-    void exerciseListReplyFinished (QNetworkReply *reply);
+    void exerciseListReplyFinished (QNetworkReply *reply, Course *course);
     void exerciseZipReplyFinished (QNetworkReply *reply, Exercise *ex);
 
 private:
-    QNetworkAccessManager manager;
+    QNetworkRequest buildRequest(QUrl url);
+    QNetworkReply* doGet(QUrl url);
+    QNetworkAccessManager *manager;
     QString accessToken;
-    Course *m_course;
 
 };
