+++ testmycode.pro
@@ -1,13 +1,16 @@
 TEMPLATE = subdirs
 
-SUBDIRS += \
+SUBDIRS = \
         src \
-        3rdparty
+        3rdparty \
+        lib/tmcclient
 
-src.depends = 3rdparty
+lib/tmcclient.depends = 3rdparty
+src.depends = lib/tmcclient 3rdparty
 
 test {
     message(Configuring test build)
     SUBDIRS += test
-    test.depends = src
+    test.depends += 3rdparty
+    test.depends += lib/tmcclient
 }
